Phase D13: Complete disability struct definitions and reconcile concurrent changes
- Add PoorLaws struct and poor_laws field to LocalLaws in geography.rs
- Add disability_severity, disabled_labor_capacity, extracted_caregivers to ClassDemographics
- Add education field to Region struct
- Add EmancipationLaw enum to laws.rs with parse_from_str method
- Add LawType::ChildLabor match arm in enact_law
- Add JusticeConfig struct with PartialEq derive to justice_system.rs
- Add missing Country field initializers in generator, save_manager, rebellions
- Add poor_laws to all 5 LocalLaws initializers in factional_domains.rs
- Add education field to Region initializers in logistics, commuting, air_cargo
- Register begging, pension, disability_config, education_config modules
- Fix CropBatch initializer missing maturity_turn and parcel_id

Iron CI/CD: build PASSED, clippy PASSED, npm build PASSED
9 pre-existing test failures in trade/religion modules (unrelated to disability work)

+++ state/src/corporate/manager.rs
@@ -2420,7 +2420,7 @@ pub fn process_mutual_aid_turn(
             // general-purpose "turns in distress" counter for mutual aid
             // circles (it's not used for furlough on non-seasonal companies).
             company.furlough_turns_accumulated += 1;
-            if company.furlough_turns_accumulated >= config.dissolution_turns as u32 {
+            if company.furlough_turns_accumulated >= config.dissolution_turns {
                 // Mark for bankruptcy via the standard corporate lifecycle.
                 // The lifecycle module will handle asset liquidation and
                 // removal from the active market.
